Moved escape.js to format.js
