+++ lib/core/format.js
