+++ lib/db/stats.js
@@ -1,5 +1,5 @@
 import {createIndex, findAll, upsertOne} from './mongo.js'
-import {encodePage, decodePage} from '../core/escape.js'
+import {encodePage, decodePage} from '../core/format.js'
 
 
 await init()
