+++ lib/page/home.js
@@ -1,4 +1,4 @@
-import {shorten} from '../core/escape.js'
+import {shorten} from '../core/format.js'
 
 
 export function createHome(latestSites) {
