+++ lib/page/stats.js
@@ -1,4 +1,4 @@
-import {shorten} from '../core/escape.js'
+import {shorten} from '../core/format.js'
 
 
 export function createStatsPage(site, stats) {
